fix/feat: added scanner to end screens. Added  texst for returning to receiver from gateway. fixed animation issue. Refactored NFC scanning logic and
feat: Added scanner to end screens. Refactored NFC logic for this. Pulled out the NFC logic as a hook and re-added it to the PacketScanner component
feat: added special text for return packet to receiver.
refactor: Extracted the Packet Description (bottom part) of the whole instruction and packet scan component, as this is required for the onboarding and offboarding screens.